fix: text color option not working
